requirements.txt
flask
streamlit
scikit-learn
pandas
numpy
matplotlib
plotly-express
joblib
